Keep a consistent coding style.
WinFile is a mess to read and never seems to adhere to a single style. This commit tries to correct that. The next step is to try to get it to build under the ReactOS Build Environment and see if this commit caused any regressions.

Also made updates to the README.md file.